Update BayesNetWrapper to return any additional number of outputs
This was updated in the predict function of the BayesNetWrapper. It now
averages also every additional return values that are returned from the
forward of the network (embeddings e.g.) and also averages them over all
samples. They are then returned at the end of the pred, aleatoric- and
epistemic-uncertainties as an unpacked list.